# docs/swagger.yaml
definitions:
...
  model.GeneralResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  model.JoinProgramModifications:
    properties:
      email:
        type: string
      modifications:
        items:
          $ref: '#/definitions/model.Modification'
        type: array
    required:
    - email
    type: object
...
  model.Modification:
    properties:
      is_set_noti:
        type: integer
      start_time:
        type: string
      time_before_notify:
        type: integer
    required:
    - start_time
    type: object
...
  model.TaskCreateOrUpdate:
    properties:
      category:
        type: string
      is_set_noti:
        default: 0
        type: integer
      link:
        type: string
      media_url:
        type: string
      start_time:
        type: string
      task_description:
        type: string
      task_name:
        maxLength: 250
        type: string
      time_before_notify:
        type: integer
    required:
    - task_name
    type: object
  model.ToDoRequest:
    properties:
      date:
        type: string
      email:
        type: string
    required:
    - date
    - email
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/tasks:
    get:
      consumes:
      - application/json
      description: Find All Tasks
      operationId: FindAllTask
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Find All Tasks
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Create new task
      operationId: SaevTask
      parameters:
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.TaskCreateOrUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Save Task
      tags:
      - Task
  /api/v1/tasks/:id:
    delete:
      consumes:
      - application/json
      description: Delete Existing Task
      operationId: DeleteTask
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Delete Task
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: Find a task by passing ID
      operationId: FindTaskById
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Find task by ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update a task by passing ID
      operationId: UpdateTask
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.TaskCreateOrUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Update Task
      tags:
      - Task
  /api/v1/tasks/done/:id:
    put:
      consumes:
      - application/json
      description: update status to 1 (done)
      operationId: UpdateDone
      parameters:
      - description: Task  ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: update status to done
      tags:
      - Task
  /api/v1/tasks/join-program/:program_id:
    post:
      consumes:
      - application/json
      description: Create tasks from program
      operationId: JoinProgram
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Modifications
        in: body
        name: modifications
        required: true
        schema:
          $ref: '#/definitions/model.JoinProgramModifications'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: join program
      tags:
      - Task
  /api/v1/tasks/plan/:program_id:
    get:
      consumes:
      - application/json
      description: List task related to that program_id for planning
      operationId: PlanTask
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: list task to plan program
      tags:
      - Task
  /api/v1/tasks/to-do:
    get:
      consumes:
      - application/json
      description: Find Task by Email and Date
      operationId: FindTaskByEmailDate
      parameters:
      - description: Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.ToDoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: Find Task by Email and Date
      tags:
      - Task
  /api/v1/tasks/un-done/:id:
    put:
      consumes:
      - application/json
      description: update status to 1 (todo)
      operationId: UpdateTodo
      parameters:
      - description: Task  ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.GeneralResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.GeneralResponse'
      summary: update status to todo
      tags:
      - Task
swagger: "2.0"